Change RoomPopup UI
