+++ plugins/love-resources/src/components/RoomPopup.svelte
@@ -44,9 +44,9 @@
   } from '@hcengineering/love'
   import { getClient } from '@hcengineering/presentation'
   import view from '@hcengineering/view'
-  import { getObjectLinkFragment } from '@hcengineering/view-resources'
+  import { DocNavLink, getObjectLinkFragment } from '@hcengineering/view-resources'
   import { toggleCamState, toggleMicState, state } from '@hcengineering/media-resources'
-  import { createEventDispatcher } from 'svelte'
+  import { createEventDispatcher, onMount } from 'svelte'
   import love from '../plugin'
   import { currentMeetingMinutes, currentRoom, infos, invites, myInfo, myOffice, myRequests, rooms } from '../stores'
   import {
@@ -64,6 +64,16 @@
   import MicSettingPopup from './meeting/MicSettingPopup.svelte'
   import RoomAccessPopup from './RoomAccessPopup.svelte'
   import ShareSettingPopup from './ShareSettingPopup.svelte'
+  import MicrophoneButton from './meeting/controls/MicrophoneButton.svelte'
+  import CameraButton from './meeting/controls/CameraButton.svelte'
+  import ShareScreenButton from './meeting/controls/ShareScreenButton.svelte'
+  import TranscriptionButton from './meeting/controls/TranscriptionButton.svelte'
+  import RoomAccessButton from './meeting/controls/RoomAccessButton.svelte'
+  import ControlBarContainer from './meeting/ControlBarContainer.svelte'
+  import SendReactionButton from './meeting/controls/SendReactionButton.svelte'
+  import RecordingButton from './meeting/controls/RecordingButton.svelte'
+  import LeaveRoomButton from './meeting/controls/LeaveRoomButton.svelte'
+  import MeetingHeader from './meeting/MeetingHeader.svelte'
 
   export let room: Room
 
@@ -79,44 +89,10 @@
   let joined: boolean = false
   $: joined = $myInfo?.room === room._id
 
-  $: isMyOffice = $myInfo?.room === $myOffice?._id
-
-  $: allowLeave = !isMyOffice && $myInfo?.room !== love.ids.Reception
-
   let info: ParticipantInfo[] = []
   $: info = $infos.filter((p) => p.room === room._id)
 
-  let allowCam: boolean = false
-  const allowShare: boolean = true
-
-  $: allowCam = $currentRoom?.type === RoomType.Video
-  $: isMicEnabled = $state.microphone?.enabled === true
-  $: isCamEnabled = $state.camera?.enabled === true
-
   const dispatch = createEventDispatcher()
-
-  async function changeShare (): Promise<void> {
-    const newValue = !$isSharingEnabled
-    const audio = newValue && $isShareWithSound
-    await liveKitClient.setScreenShareEnabled(newValue, audio)
-    dispatch('close')
-  }
-
-  async function leave (): Promise<void> {
-    await leaveRoom($myInfo, $myOffice)
-    dispatch('close')
-  }
-
-  async function end (): Promise<void> {
-    if (isOffice(room) && $myInfo !== undefined) {
-      await endMeeting(room, $rooms, $infos, $myInfo)
-    }
-    dispatch('close')
-  }
-
-  function shareSettings (e: MouseEvent): void {
-    showPopup(ShareSettingPopup, {}, eventToHTMLElement(e))
-  }
 
   async function connect (): Promise<void> {
     await tryConnect($myInfo, room, info, $myRequests, $invites)
@@ -139,23 +115,6 @@
     }
   }
 
-  function micSettings (e: MouseEvent): void {
-    showPopup(MicSettingPopup, {}, eventToHTMLElement(e))
-  }
-
-  function camSettings (e: MouseEvent): void {
-    showPopup(CamSettingPopup, {}, eventToHTMLElement(e))
-  }
-
-  function setAccess (e: MouseEvent): void {
-    showPopup(RoomAccessPopup, { room }, eventToHTMLElement(e), (res) => {
-      if (res !== undefined) {
-        room.access = res
-      }
-    })
-  }
-
-  const me = getCurrentEmployee()
   function canGoBack (joined: boolean, location: Location, meetingMinutes?: MeetingMinutes): boolean {
     if (!joined) return false
     if (location.path[2] !== loveId) return true
@@ -168,13 +127,8 @@
   }
 </script>
 
-<div class="antiPopup room-popup">
-  <div class="room-label"><Label label={love.string.Room} /></div>
-  <div class="title overflow-label">
-    {#await getRoomName(room) then name}
-      {name}
-    {/await}
-  </div>
+<div class="antiPopup room-popup flex-gap-4">
+  <MeetingHeader {room} />
   <div class="room-popup__content">
     <Scroller padding={'0.5rem'} stickedScrollBars>
       <div class="room-popup__content-grid">
@@ -188,98 +142,37 @@
       </div>
     </Scroller>
   </div>
-  {#if joined && $lkSessionConnected}
-    <div class="room-btns" class:no-video={!allowCam}>
-      <SplitButton
-        size={'large'}
-        icon={isMicEnabled ? love.icon.MicEnabled : love.icon.MicDisabled}
-        label={isMicEnabled ? love.string.Mute : love.string.UnMute}
-        showTooltip={{
-          label: isMicEnabled ? love.string.Mute : love.string.UnMute
-        }}
-        action={toggleMicState}
-        secondIcon={IconUpOutline}
-        secondAction={micSettings}
-        separate
-      />
-      {#if allowCam}
-        <SplitButton
-          size={'large'}
-          icon={isCamEnabled ? love.icon.CamEnabled : love.icon.CamDisabled}
-          label={isCamEnabled ? love.string.StopVideo : love.string.StartVideo}
-          showTooltip={{
-            label: isCamEnabled ? love.string.StopVideo : love.string.StartVideo
-          }}
-          action={toggleCamState}
-          secondIcon={IconUpOutline}
-          secondAction={camSettings}
-          separate
-        />
+    <div class="flex-between gap-2">
+      {#if joined && $lkSessionConnected}
+        <div class="flex-between gap-2">
+          <MicrophoneButton size="medium" />
+          <CameraButton size="medium" />
+          <ShareScreenButton size="medium" on:changeShare={() => dispatch('close')} />
+        </div>
       {/if}
-      {#if allowShare}
-        <SplitButton
-          size={'large'}
-          icon={$isSharingEnabled ? love.icon.SharingEnabled : love.icon.SharingDisabled}
-          showTooltip={{ label: $isSharingEnabled ? love.string.StopShare : love.string.Share }}
-          disabled={($screenSharing && !$isSharingEnabled) || !$lkSessionConnected}
-          action={changeShare}
-          secondIcon={IconUpOutline}
-          secondAction={shareSettings}
-          separate
-        />
-      {/if}
-      <ModernButton
-        icon={roomAccessIcon[room.access]}
-        label={roomAccessLabel[room.access]}
-        tooltip={{ label: love.string.ChangeAccess }}
-        kind={'secondary'}
-        size={'large'}
-        disabled={isOffice(room) && room.person !== me}
-        on:click={setAccess}
-      />
-    </div>
-  {/if}
-  {#if $location.path[2] !== loveId || (joined && (allowLeave || isMyOffice)) || !joined}
-    <div class="btns flex-row-center flex-reverse flex-no-shrink w-full flex-gap-2">
-      {#if joined}
-        {#if allowLeave}
-          <ModernButton
-            label={love.string.LeaveRoom}
-            icon={love.icon.LeaveRoom}
-            size={'large'}
-            kind={'negative'}
-            on:click={leave}
-          />
-        {:else if isMyOffice}
-          <ModernButton
-            label={love.string.EndMeeting}
-            icon={love.icon.LeaveRoom}
-            size={'large'}
-            kind={'negative'}
-            on:click={end}
-          />
-        {/if}
-      {:else}
-        <ModernButton
-          icon={love.icon.EnterRoom}
-          label={love.string.EnterRoom}
-          size={'large'}
-          kind={'primary'}
-          autoFocus
-          on:click={connect}
-        />
-      {/if}
+      <div style="width: auto" />
       {#if canGoBack(joined, $location, $currentMeetingMinutes)}
         <ModernButton
           icon={IconArrowLeft}
           label={love.string.MeetingMinutes}
           kind={'primary'}
-          size={'large'}
+          size={'medium'}
           on:click={back}
         />
       {/if}
+      {#if joined}
+        <LeaveRoomButton {room} noLabel={false} size="medium" on:leave={() => dispatch('close')} />
+      {:else}
+        <ModernButton
+          icon={love.icon.EnterRoom}
+          label={love.string.EnterRoom}
+          size={'medium'}
+          kind={'primary'}
+          autoFocus
+          on:click={connect}
+        />
+      {/if}
     </div>
-  {/if}
 </div>
 
 <style lang="scss">
