+++ plugins/love-resources/src/components/meeting/ControlBarContainer.svelte
