+++ plugins/love-resources/src/components/meeting/widget/ChatTab.svelte
@@ -13,15 +13,13 @@
 // limitations under the License.
 -->
 <script lang="ts">
-  import love, { MeetingMinutes, Room } from '@hcengineering/love'
+  import love, { MeetingMinutes } from '@hcengineering/love'
   import { ChannelEmbeddedContent } from '@hcengineering/chunter-resources'
   import { ActivityMessage } from '@hcengineering/activity'
   import { updateTabData, WidgetState } from '@hcengineering/workbench-resources'
-  import MeetingWidgetHeader from './MeetingWidgetHeader.svelte'
 
   export let widgetState: WidgetState
   export let meetingMinutes: MeetingMinutes
-  export let room: Room
   export let height: string
   export let width: string
 
@@ -44,7 +42,4 @@
   onReply={replyToThread}
   on:close
 >
-  <svelte:fragment slot="header">
-    <MeetingWidgetHeader doc={meetingMinutes} {room} on:close />
-  </svelte:fragment>
 </ChannelEmbeddedContent>
