+++ plugins/love-resources/src/components/meeting/widget/MeetingWidget.svelte
@@ -28,8 +28,9 @@
   import RoomAccessButton from '../controls/RoomAccessButton.svelte'
   import RecordingButton from '../controls/RecordingButton.svelte'
   import TranscriptionButton from '../controls/TranscriptionButton.svelte'
-  import ControlBarContainer from '../../ControlBarContainer.svelte'
+  import ControlBarContainer from '../ControlBarContainer.svelte'
   import SendReactionButton from '../controls/SendReactionButton.svelte'
+  import MeetingWidgetHeader from './MeetingWidgetHeader.svelte'
 
   export let widgetState: WidgetState | undefined
   export let height: string
@@ -41,6 +42,8 @@
   let isMeetingMinutesLoaded = false
 
   let room: Room | undefined = undefined
+
+  let contentHeight: number = 0
 
   $: room = $currentRoom
 
@@ -76,22 +79,27 @@
   }
 </script>
 
-{#if widgetState && room}
-  {#if widgetState.tab === 'video'}
-    <VideoTab {room} doc={meetingMinutes} on:close={handleClose} />
-  {:else if widgetState.tab === 'chat'}
-    {#if !isMeetingMinutesLoaded}
-      <Loading />
-    {:else if meetingMinutes}
-      <ChatTab {meetingMinutes} {room} {widgetState} {height} {width} on:close={handleClose} />
+{#if widgetState !== undefined && room}
+  <div>
+    <MeetingWidgetHeader doc={meetingMinutes} {room} on:close={handleClose} />
+  </div>
+  <div style="height: 100%;" bind:clientHeight={contentHeight}>
+    {#if widgetState.tab === 'video'}
+      <VideoTab {room} doc={meetingMinutes} on:close={handleClose} />
+    {:else if widgetState.tab === 'chat'}
+      {#if !isMeetingMinutesLoaded}
+        <Loading />
+      {:else if meetingMinutes}
+        <ChatTab {meetingMinutes} {widgetState} height={contentHeight + 'px'} {width} on:close={handleClose} />
+      {/if}
+    {:else if widgetState.tab === 'transcription'}
+      {#if !isMeetingMinutesLoaded}
+        <Loading />
+      {:else if meetingMinutes}
+        <TranscriptionTab {meetingMinutes} {room} {widgetState} height={contentHeight + 'px'} {width} on:close={handleClose} />
+      {/if}
     {/if}
-  {:else if widgetState.tab === 'transcription'}
-    {#if !isMeetingMinutesLoaded}
-      <Loading />
-    {:else if meetingMinutes}
-      <TranscriptionTab {meetingMinutes} {room} {widgetState} {height} {width} on:close={handleClose} />
-    {/if}
-  {/if}
+  </div>
   <ControlBarContainer>
     <svelte:fragment slot="right">
       <RoomAccessButton {room} size="small"/>
@@ -102,7 +110,7 @@
       <TranscriptionButton {room} size="small"/>
     </svelte:fragment>
     <svelte:fragment slot="left">
-      <LeaveRoomButton noLabel={true} size="small" />
+      <LeaveRoomButton {room} noLabel={true} size="small" />
     </svelte:fragment>
   </ControlBarContainer>
 {/if}
