+++ plugins/love-resources/src/components/meeting/widget/MeetingWidgetHeader.svelte
@@ -29,10 +29,10 @@
   import RoomModal from '../../RoomModal.svelte'
   import { currentRoom } from '../../../stores'
   import { screenSharing } from '../../../utils'
-  import ControlBar from '../../ControlBar.svelte'
+  import ControlBar from '../ControlBar.svelte'
   import MeetingOptionsButton from '../controls/MeetingOptionsButton.svelte'
   import SendReactionButton from '../controls/SendReactionButton.svelte'
-  import ControlBarContainer from '../../ControlBarContainer.svelte'
+  import ControlBarContainer from '../ControlBarContainer.svelte'
   import RoomAccessButton from '../controls/RoomAccessButton.svelte'
   import LeaveRoomButton from '../controls/LeaveRoomButton.svelte'
   import RecordingButton from '../controls/RecordingButton.svelte'
