+++ plugins/love-resources/src/index.ts
@@ -2,7 +2,7 @@ import { getMetadata, type Resources } from '@hcengineering/platform'
 import aiBot from '@hcengineering/ai-bot'
 import { AccountRole, getCurrentAccount, hasAccountRole } from '@hcengineering/core'
 
-import ControlExt from './components/ControlExt.svelte'
+import ControlExt from './components/meeting/ControlExt.svelte'
 import EditMeetingData from './components/EditMeetingData.svelte'
 import Main from './components/Main.svelte'
 import MeetingData from './components/MeetingData.svelte'
